# swagger.json
{"openapi":"3.0.0","paths":{"/healthcheck":{"get":{"operationId":"HealthcheckController_getStatus","summary":"Get server status","parameters":[],"responses":{"200":{"description":"ok","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthcheckDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["healthcheck"]}},"/session/login":{"post":{"operationId":"SessionController_postLogin","summary":"Post the login credentials data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the jwt data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostLoginResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["session"]}},"/session/register":{"post":{"operationId":"SessionController_postCredentials","summary":"Post the credentials data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostCredentialsReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the credentials data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["session"]}},"/user/info":{"get":{"operationId":"UserController_getUser","summary":"Get the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/alter":{"put":{"operationId":"UserController_putUser","summary":"Put the user data","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PutUserReqDto"}}}},"responses":{"200":{"description":"Returns a JSON with the user data","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}},"/user/remove":{"delete":{"operationId":"UserController_deleteUser","summary":"Delete the user data","parameters":[],"responses":{"200":{"description":"Returns a JSON with the user status","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeleteUserResDto"}}}},"500":{"description":"Internal server error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ErrorDto"}}}}},"tags":["user"],"security":[{"auth":[]}]}}},"info":{"title":"pdp-ms-user","description":"cosmetics e-commerce","version":"1.0.0","contact":{}},"tags":[],"servers":[{"url":"/pdp-ms-user"}],"components":{"securitySchemes":{"auth":{"scheme":"bearer","bearerFormat":"JWT","type":"http","name":"JWT","in":"header"}},"schemas":{"HealthcheckDto":{"type":"object","properties":{"name":{"type":"string"},"version":{"type":"string"}},"required":["name","version"]},"ErrorDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]},"PostLoginReqDto":{"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}},"required":["email","password"]},"PostLoginResDto":{"type":"object","properties":{"jwt":{"type":"string"}},"required":["jwt"]},"PostCredentialsReqDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"password":{"type":"string"},"is_admin":{"type":"boolean"}},"required":["name","email","password","is_admin"]},"UserDto":{"type":"object","properties":{"name":{"type":"string"},"email":{"type":"string"},"is_admin":{"type":"boolean"}},"required":["name","email","is_admin"]},"GetUserResDto":{"type":"object","properties":{"user":{"$ref":"#/components/schemas/UserDto"}},"required":["user"]},"PutUserReqDto":{"type":"object","properties":{"name":{"type":"string"},"password":{"type":"string"}},"required":["name","password"]},"DeleteUserResDto":{"type":"object","properties":{"statusCode":{"type":"number"},"message":{"type":"string"}},"required":["statusCode","message"]}}}}